# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Server {
  id: string;
  name: string;
  ownerId: string;
  inviteCode: string;
  memberCount: number;
  createdAt: string;
}

export interface CreateServerBody {
  name: string;
  ownerId: string;
  ownerName: string;
  ownerColor?: string;
}

export interface UpdateServerBody {
  name: string;
  requesterId: string;
}

export interface OwnerActionBody {
  requesterId: string;
}

export interface JoinServerBody {
  userId: string;
  userName: string;
  userColor?: string;
  inviteCode?: string;
}

export interface JoinByCodeBody {
  userId: string;
  userName: string;
  userColor?: string;
  inviteCode: string;
}

export interface FindByCodeBody {
  inviteCode: string;
}

export interface InviteCodeResponse {
  inviteCode: string;
}

export interface TransferCodeResponse {
  transferCode: string;
}

export interface RedeemTransferBody {
  userId: string;
  userName: string;
  transferCode: string;
}

export interface Message {
  id: string;
  serverId: string;
  userId: string;
  userName: string;
  userColor?: string;
  content: string;
  createdAt: string;
}

export interface CreateMessageBody {
  userId: string;
  userName: string;
  userColor?: string;
  content: string;
}

export interface DeleteMessageBody {
  requesterId: string;
}

export interface PresenceBody {
  userId: string;
  userName: string;
  userColor?: string;
}

export interface PresenceResponse {
  onlineCount: number;
}

export interface OnlineUser {
  userId: string;
  userName: string;
  userColor?: string;
  lastSeen: string;
}

export interface TypingBody {
  userId: string;
  userName: string;
  isTyping: boolean;
}

export interface TypingResponse {
  success: boolean;
}

export interface TypingUser {
  userId: string;
  userName: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListMessagesParams = {
  limit?: number;
  before?: string;
};