# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chat Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers/{serverId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /servers/{serverId}/messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (owner or message author)
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteMessageBody"
      responses:
        "204":
          description: Message deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{serverId}/presence:
    post:
      operationId: updatePresence
      tags: [presence]
      summary: Update user presence (heartbeat)
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenceBody"
      responses:
        "200":
          description: Presence updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresenceResponse"
    get:
      operationId: getOnlineUsers
      tags: [presence]
      summary: Get online users for a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of online users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnlineUser"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
        serverId:
          type: string
        userId:
          type: string
        userName:
          type: string
        userColor:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - userId
        - userName
        - content
        - createdAt
    CreateMessageBody:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        userColor:
          type: string
        content:
          type: string
      required:
        - userId
        - userName
        - content
    DeleteMessageBody:
      type: object
      properties:
        requesterId:
          type: string
      required:
        - requesterId
    PresenceBody:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        userColor:
          type: string
      required:
        - userId
        - userName
    PresenceResponse:
      type: object
      properties:
        onlineCount:
          type: integer
      required:
        - onlineCount
    OnlineUser:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        userColor:
          type: string
        lastSeen:
          type: string
          format: date-time
      required:
        - userId
        - userName
        - lastSeen
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error