# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all servers
 */
export const ListServersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListServersResponse = zod.array(ListServersResponseItem);

/**
 * @summary Create a new server
 */
export const CreateServerBody = zod.object({
  name: zod.string(),
  ownerId: zod.string(),
  ownerName: zod.string(),
  ownerColor: zod.string().optional(),
});

/**
 * @summary Get a server by ID
 */
export const GetServerParams = zod.object({
  serverId: zod.coerce.string(),
});

export const GetServerResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update server (owner only)
 */
export const UpdateServerParams = zod.object({
  serverId: zod.coerce.string(),
});

export const UpdateServerBody = zod.object({
  name: zod.string(),
  requesterId: zod.string(),
});

export const UpdateServerResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete server (owner only)
 */
export const DeleteServerParams = zod.object({
  serverId: zod.coerce.string(),
});

export const DeleteServerBody = zod.object({
  requesterId: zod.string(),
});

/**
 * @summary Join a server using invite code or server ID
 */
export const JoinServerParams = zod.object({
  serverId: zod.coerce.string(),
});

export const JoinServerBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
  inviteCode: zod.string().optional(),
});

export const JoinServerResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Generate a new invite code (owner only)
 */
export const GenerateInviteCodeParams = zod.object({
  serverId: zod.coerce.string(),
});

export const GenerateInviteCodeBody = zod.object({
  requesterId: zod.string(),
});

export const GenerateInviteCodeResponse = zod.object({
  inviteCode: zod.string(),
});

/**
 * @summary Generate an ownership transfer code (owner only)
 */
export const GenerateTransferCodeParams = zod.object({
  serverId: zod.coerce.string(),
});

export const GenerateTransferCodeBody = zod.object({
  requesterId: zod.string(),
});

export const GenerateTransferCodeResponse = zod.object({
  transferCode: zod.string(),
});

/**
 * @summary Redeem an ownership transfer code
 */
export const RedeemTransferCodeParams = zod.object({
  serverId: zod.coerce.string(),
});

export const RedeemTransferCodeBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  transferCode: zod.string(),
});

export const RedeemTransferCodeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Join a server using invite code
 */
export const JoinByInviteCodeParams = zod.object({
  serverId: zod.coerce.string(),
});

export const JoinByInviteCodeBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
  inviteCode: zod.string(),
});

export const JoinByInviteCodeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Find a server by invite code
 */
export const FindServerByCodeBody = zod.object({
  inviteCode: zod.string(),
});

export const FindServerByCodeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  ownerId: zod.string(),
  inviteCode: zod.string(),
  memberCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List messages for a server
 */
export const ListMessagesParams = zod.object({
  serverId: zod.coerce.string(),
});

export const listMessagesQueryLimitDefault = 50;

export const ListMessagesQueryParams = zod.object({
  limit: zod.coerce.number().default(listMessagesQueryLimitDefault),
  before: zod.coerce.string().optional(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.string(),
  serverId: zod.string(),
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a message
 */
export const CreateMessageParams = zod.object({
  serverId: zod.coerce.string(),
});

export const CreateMessageBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
  content: zod.string(),
});

/**
 * @summary Delete a message (owner or message author)
 */
export const DeleteMessageParams = zod.object({
  serverId: zod.coerce.string(),
  messageId: zod.coerce.string(),
});

export const DeleteMessageBody = zod.object({
  requesterId: zod.string(),
});

/**
 * @summary Update user presence (heartbeat)
 */
export const UpdatePresenceParams = zod.object({
  serverId: zod.coerce.string(),
});

export const UpdatePresenceBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
});

export const UpdatePresenceResponse = zod.object({
  onlineCount: zod.number(),
});

/**
 * @summary Get online users for a server
 */
export const GetOnlineUsersParams = zod.object({
  serverId: zod.coerce.string(),
});

export const GetOnlineUsersResponseItem = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  userColor: zod.string().optional(),
  lastSeen: zod.coerce.date(),
});
export const GetOnlineUsersResponse = zod.array(GetOnlineUsersResponseItem);

/**
 * @summary Send typing indicator
 */
export const SendTypingIndicatorParams = zod.object({
  serverId: zod.coerce.string(),
});

export const SendTypingIndicatorBody = zod.object({
  userId: zod.string(),
  userName: zod.string(),
  isTyping: zod.boolean(),
});

export const SendTypingIndicatorResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get currently typing users
 */
export const GetTypingUsersParams = zod.object({
  serverId: zod.coerce.string(),
});

export const GetTypingUsersResponseItem = zod.object({
  userId: zod.string(),
  userName: zod.string(),
});
export const GetTypingUsersResponse = zod.array(GetTypingUsersResponseItem);